+++ src/main/resources/static/css/mypage/mypage.css
@@ -212,15 +212,15 @@ aside.show {
 
 }
 .pwdfont{
-    width: 455px;
+    width: 600px;
     height: 78px;
     font-family: 'Cafe24SsurroundAir';
     font-style: normal;
     font-weight: 300;
     font-size: 30px;
     line-height: 39px;
-    margin-left: 150px;
-    margin-top: 40px;
+    margin-left: 100px;
+    margin-top: 50px;
     margin-bottom: 40px;
     color: #FFFFFF;
