+++ src/main/resources/templates/main/main.html
@@ -5,36 +5,37 @@
     <title>남호정</title>
     <link rel="stylesheet" th:href="@{/css/reset.css}">
     <link rel="stylesheet" th:href="@{/css/mainpage/mainpage-darkmode.css}">
+    <link rel="stylesheet" th:href="@{/css/mypage/mypage.css}">
     <!--    <link rel="stylesheet" th:href="@{/css/mainpage/mainpage-lightmode.css}">-->
-    <style>
-        /* 손전등 */
-        :root {
-            --cursorX: 70vw;
-            --cursorY: 70vh;
-        }
-        :root:before {
-            content: '';
-            display: block;
-            width: 100%;
-            height: 100%;
-            position: fixed;
-            pointer-events: none;
-            z-index: 999999;
-            background: radial-gradient(
-                    circle 15vmax at var(--cursorX) var(--cursorY),
-                    rgba(0,0,0,0) 0%,
-                    rgba(0,0,0,.5) 80%,
-                    rgba(0,0,0,.95) 100%)
-        }
+<!--    <style>-->
+<!--        /* 손전등 */-->
+<!--        :root {-->
+<!--            &#45;&#45;cursorX: 70vw;-->
+<!--            &#45;&#45;cursorY: 70vh;-->
+<!--        }-->
+<!--        :root:before {-->
+<!--            content: '';-->
+<!--            display: block;-->
+<!--            width: 100%;-->
+<!--            height: 100%;-->
+<!--            position: fixed;-->
+<!--            pointer-events: none;-->
+<!--            z-index: 999999;-->
+<!--            background: radial-gradient(-->
+<!--                    circle 15vmax at var(&#45;&#45;cursorX) var(&#45;&#45;cursorY),-->
+<!--                    rgba(0,0,0,0) 0%,-->
+<!--                    rgba(0,0,0,.5) 80%,-->
+<!--                    rgba(0,0,0,.95) 100%)-->
+<!--        }-->
 
-        * {
-            box-sizing: border-box;
-            text-rendering: optimizeLegibility;
-            -webkit-font-smoothing: antialiased;
-            -moz-osx-font-smoothing: grayscale;
-            font-kerning: auto;
-        }
-    </style>
+<!--        * {-->
+<!--            box-sizing: border-box;-->
+<!--            text-rendering: optimizeLegibility;-->
+<!--            -webkit-font-smoothing: antialiased;-->
+<!--            -moz-osx-font-smoothing: grayscale;-->
+<!--            font-kerning: auto;-->
+<!--        }-->
+<!--    </style>-->
 </head>
 <body>
 
@@ -47,13 +48,45 @@
         <span onclick="location.href='/notice'">공지사항</span>
         <span onclick="location.href='/post'">게시판</span>
         <span>MBTI 검사</span>
-        <span>마이페이지</span>
+        <span id="mypageBtn">마이페이지</span>
         <span>로그인</span>
         <span>회원가입</span>
     </nav>
 
 </header>
 
+
+<aside id="myPageAside">
+    <p class="mypagefont">마이페이지</p>
+    <img th:src="@{/images/mypageimages/profile.png}" alt="프로필 사진" class="profile-photo" id="profilePhoto">
+    <p id="nickname" class="nicknamefont">닉네임 : <span th:text="${myPageDTO.userName}"></span></p>
+    <p id="rank" class="rankfont">등급: 어쩌구</p>
+    <p class="infofont" id="info">개인 정보</p>
+    <p class="activefont" id="active">활동 기록</p>
+    <p class="MBTIfont">MBTI</p>
+    <p class="logoutfont">로그아웃</p>
+</aside>
+<div class="rankpopup" id="rank-popup">
+    <p class="scorefont">활동 점수 : </p>
+    <p class="nextfont">다음 등급 : </p>
+    <p class="needfont">필요 점수 : </p>
+</div>
+
+<div class="modal-background" id="modal-background">
+    <div class="pwdpopup" id="pwd-popup">
+        <p class="pwdfont">비밀번호를 입력하여 본인임을 확인해 주세요!</p>
+        <div class="pwd-container">
+            <span class="pwdfont2">비밀번호</span> <input type="password" id="pwdInput" placeholder="비밀번호">
+            <span class="toggle-password" id="togglePassword">&#128065;</span>
+        </div>
+
+        <div class="popup-buttons">
+            <button id="closePopupBtn" class="cancle-btn">취소</button>
+            <button id="submitPasswordBtn" class="confirm-btn">확인</button>
+        </div>
+
+    </div>
+</div>
 <main>
     <!--메인 1페이지-->
     <div class="main1">
@@ -121,6 +154,85 @@
 <footer>
     <div class="footer"><p>ⓒ 2024 MergePing. All right reserved.  </p></div>
 </footer>
+<script>
+    // JavaScript로 토글 기능 추가
+    const mypageBtn = document.getElementById("mypageBtn");
+    const myPageAside = document.getElementById("myPageAside");
+
+    // 버튼 클릭 시 aside 창 표시/숨김 전환
+    mypageBtn.addEventListener("click", () => {
+        if (myPageAside.classList.contains("show")) {
+            // 보이는 상태라면 숨기기
+            myPageAside.classList.remove("show");
+            setTimeout(() => {
+                myPageAside.style.display = "none"; // 애니메이션 후 display를 none으로
+            }, 400); // transition과 동일한 시간 설정
+        } else {
+            // 숨겨진 상태라면 보이게 하기
+            myPageAside.style.display = "block"; // 먼저 display를 block으로 설정
+            setTimeout(() => {
+                myPageAside.classList.add("show");
+            }, 10); // 약간의 지연 시간 후에 show 클래스 추가
+        }
+    });
+
+    const gradeElement = document.getElementById('rank');
+    const popup = document.getElementById('rank-popup');
+
+    gradeElement.addEventListener('mouseenter', () => {
+        popup.style.display = 'block'; // 팝업 표시
+    });
+
+    gradeElement.addEventListener('mouseleave', () => {
+        popup.style.display = 'none'; // 팝업 숨기기
+    });
+
+    const personalInfoBtn = document.getElementById('info');
+    const passwordPopup = document.getElementById('pwd-popup');
+    const closePopupBtn = document.getElementById('closePopupBtn');
+    const togglePassword = document.getElementById('togglePassword');
+    const passwordInput = document.getElementById('pwdInput');
+    const submitPasswordBtn = document.getElementById('submitPasswordBtn');
+    const modal = document.querySelector('.modal-background');
+
+    //모달
+    personalInfoBtn.addEventListener('click', () => {
+        passwordPopup.style.display = 'block';// 비밀번호 팝업 표시
+        document.body.style.overflow = "hidden"; // 스크롤 비활성화
+    });
+
+    personalInfoBtn.addEventListener("click", ()=>{
+        modal.style.display="flex";
+        if (modal.style.display === "flex") {
+            closePopupBtn.addEventListener("click", () => {
+                modal.style.display = "none";
+                document.body.style.overflow = "auto"; // 스크롤 다시 활성화
+            });
+        }
+    });
+
+
+
+    // 비밀번호 보기/숨기기 토글
+    togglePassword.addEventListener('click', () => {
+        const type = passwordInput.getAttribute('type') === 'password' ? 'text' : 'password';
+        passwordInput.setAttribute('type', type);
+        togglePassword.textContent = type === 'password' ? '👁️' : '🙈'; // 아이콘 변경
+    });
+    // 확인 버튼 클릭 시 다른 URL로 이동
+    submitPasswordBtn.addEventListener('click', () => {
+        // URL을 원하는 주소로 변경하세요
+        location.href = '/userinfo';
+    });
+
+    const activebtn = document.getElementById('active');
+    activebtn.addEventListener('click', () => {
+        // URL을 원하는 주소로 변경하세요
+        location.href = '/useractive';
+    });
+
+
+</script>
 
 <script>
     const floatBtn1 = document.getElementById("light");
